Add integration component and implement the ProductCompositeServiceImpl closes #79

+++ spring_boot/api/src/main/java/in/ushatech/api/core/model/Product.java
@@ -1,5 +1,5 @@
 package in.ushatech.api.core.model;
 
-public record Product(int productId, String name, double weight, String serviceAddress)
+public record Product(int productId, String name, int weight, String serviceAddress)
 {
 }
